pyproject.toml
[tool.poetry]
name = "eventfeed"
version = "0.0.1"
description = ""
authors = ["Einenlum <yann.rabiller@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
djangorestframework = "^3.11.1"
drf-yasg = "^1.17.1"
django-guardian = "^2.3.0"
django-environ = "^0.4.5"
psycopg2 = "^2.8.6"
gunicorn = "^20.0.4"
django-heroku = "^0.3.1"
mailjet_rest = "^1.3.3"
celery = "^4.4.7"
redis = "^3.5.3"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pytest-django = "^3.9.0"
pytest-xdist = "^2.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
